+++ src/driver/controller-client-common/index.js
@@ -1,3 +1,4 @@
+const _ = require("lodash");
 const { CsiBaseDriver } = require("../index");
 const { GrpcError, grpc } = require("../../utils/grpc");
 const cp = require("child_process");
@@ -41,7 +42,7 @@ class ControllerClientCommonDriver extends CsiBaseDriver {
 
       options.service.identity.capabilities.volume_expansion = [
         //"UNKNOWN",
-        "ONLINE",
+        //"ONLINE",
         //"OFFLINE"
       ];
     }
@@ -514,6 +515,11 @@ class ControllerClientCommonDriver extends CsiBaseDriver {
         driver.options.instance_id;
     }
 
+    let accessible_topology;
+    if (typeof this.getAccessibleTopology === "function") {
+      accessible_topology = await this.getAccessibleTopology();
+    }
+
     const res = {
       volume: {
         volume_id: name,
@@ -521,6 +527,7 @@ class ControllerClientCommonDriver extends CsiBaseDriver {
         capacity_bytes: 0,
         content_source: volume_content_source,
         volume_context,
+        accessible_topology,
       },
     };
 
@@ -570,14 +577,20 @@ class ControllerClientCommonDriver extends CsiBaseDriver {
    * @param {*} call
    */
   async GetCapacity(call) {
-    // really capacity is not used at all with nfs in this fashion, so no reason to enable
-    // here even though it is technically feasible.
-    throw new GrpcError(
-      grpc.status.UNIMPLEMENTED,
-      `operation not supported by driver`
-    );
-
     const driver = this;
+
+    if (
+      !driver.options.service.controller.capabilities.rpc.includes(
+        "GET_CAPACITY"
+      )
+    ) {
+      // really capacity is not used at all with nfs in this fashion, so no reason to enable
+      // here even though it is technically feasible.
+      throw new GrpcError(
+        grpc.status.UNIMPLEMENTED,
+        `operation not supported by driver`
+      );
+    }
 
     if (call.request.volume_capabilities) {
       const result = this.assertCapabilities(call.request.volume_capabilities);