+++ src/driver/controller-zfs/index.js
@@ -1,10 +1,7 @@
 const { CsiBaseDriver } = require("../index");
-const SshClient = require("../../utils/ssh").SshClient;
 const { GrpcError, grpc } = require("../../utils/grpc");
 const sleep = require("../../utils/general").sleep;
 const getLargestNumber = require("../../utils/general").getLargestNumber;
-
-const { Zetabyte, ZfsSshProcessManager } = require("../../utils/zfs");
 
 const Handlebars = require("handlebars");
 const uuidv4 = require("uuid").v4;
@@ -32,14 +29,20 @@ const VOLUME_CONTEXT_PROVISIONER_INSTANCE_ID_PROPERTY_NAME =
   "democratic-csi:volume_context_provisioner_instance_id";
 
 /**
- * Base driver to provisin zfs assets over ssh.
+ * Base driver to provisin zfs assets using zfs cli commands.
  * Derived drivers only need to implement:
+ *  - getExecClient()
+ *  - async getZetabyte()
+ *  - async setZetabyteCustomOptions(options) // optional
  *  - getDriverZfsResourceType() // return "filesystem" or "volume"
+ *  - getFSTypes() // optional
+ *  - getAccessModes() // optional
+ *  - async getAccessibleTopology() // optional
  *  - async createShare(call, datasetName) // return appropriate volume_context for Node operations
  *  - async deleteShare(call, datasetName) // no return expected
  *  - async expandVolume(call, datasetName) // no return expected, used for restarting services etc if needed
  */
-class ControllerZfsSshBaseDriver extends CsiBaseDriver {
+class ControllerZfsBaseDriver extends CsiBaseDriver {
   constructor(ctx, options) {
     super(...arguments);
 
@@ -146,42 +149,6 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
     }
   }
 
-  getSshClient() {
-    return new SshClient({
-      logger: this.ctx.logger,
-      connection: this.options.sshConnection,
-    });
-  }
-
-  async getZetabyte() {
-    const sshClient = this.getSshClient();
-    const options = {};
-    options.executor = new ZfsSshProcessManager(sshClient);
-    options.idempotent = true;
-
-    if (
-      this.options.zfs.hasOwnProperty("cli") &&
-      this.options.zfs.cli &&
-      this.options.zfs.cli.hasOwnProperty("paths")
-    ) {
-      options.paths = this.options.zfs.cli.paths;
-    }
-
-    if (
-      this.options.zfs.hasOwnProperty("cli") &&
-      this.options.zfs.cli &&
-      this.options.zfs.cli.hasOwnProperty("sudoEnabled")
-    ) {
-      options.sudo = this.getSudoEnabled();
-    }
-
-    if (typeof this.setZetabyteCustomOptions === "function") {
-      await this.setZetabyteCustomOptions(options);
-    }
-
-    return new Zetabyte(options);
-  }
-
   getSudoEnabled() {
     return this.options.zfs.cli && this.options.zfs.cli.sudoEnabled === true;
   }
@@ -218,6 +185,43 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
     await zb.zfs.destroy(datasetName + "@%", options);
   }
 
+  getFSTypes() {
+    const driverZfsResourceType = this.getDriverZfsResourceType();
+    switch (driverZfsResourceType) {
+      case "filesystem":
+        return ["nfs", "cifs"];
+      case "volume":
+        return ["ext3", "ext4", "ext4dev", "xfs"];
+    }
+  }
+
+  getAccessModes() {
+    const driverZfsResourceType = this.getDriverZfsResourceType();
+    switch (driverZfsResourceType) {
+      case "filesystem":
+        return [
+          "UNKNOWN",
+          "SINGLE_NODE_WRITER",
+          "SINGLE_NODE_SINGLE_WRITER", // added in v1.5.0
+          "SINGLE_NODE_MULTI_WRITER", // added in v1.5.0
+          "SINGLE_NODE_READER_ONLY",
+          "MULTI_NODE_READER_ONLY",
+          "MULTI_NODE_SINGLE_WRITER",
+          "MULTI_NODE_MULTI_WRITER",
+        ];
+      case "volume":
+        return [
+          "UNKNOWN",
+          "SINGLE_NODE_WRITER",
+          "SINGLE_NODE_SINGLE_WRITER", // added in v1.5.0
+          "SINGLE_NODE_MULTI_WRITER", // added in v1.5.0
+          "SINGLE_NODE_READER_ONLY",
+          "MULTI_NODE_READER_ONLY",
+          "MULTI_NODE_SINGLE_WRITER",
+        ];
+    }
+  }
+
   assertCapabilities(capabilities) {
     const driverZfsResourceType = this.getDriverZfsResourceType();
     this.ctx.logger.verbose("validating capabilities: %j", capabilities);
@@ -234,24 +238,13 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
 
           if (
             capability.mount.fs_type &&
-            !["nfs", "cifs"].includes(capability.mount.fs_type)
+            !this.getFSTypes().includes(capability.mount.fs_type)
           ) {
             message = `invalid fs_type ${capability.mount.fs_type}`;
             return false;
           }
 
-          if (
-            ![
-              "UNKNOWN",
-              "SINGLE_NODE_WRITER",
-              "SINGLE_NODE_SINGLE_WRITER", // added in v1.5.0
-              "SINGLE_NODE_MULTI_WRITER", // added in v1.5.0
-              "SINGLE_NODE_READER_ONLY",
-              "MULTI_NODE_READER_ONLY",
-              "MULTI_NODE_SINGLE_WRITER",
-              "MULTI_NODE_MULTI_WRITER",
-            ].includes(capability.access_mode.mode)
-          ) {
+          if (!this.getAccessModes().includes(capability.access_mode.mode)) {
             message = `invalid access_mode, ${capability.access_mode.mode}`;
             return false;
           }
@@ -261,26 +254,14 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
           if (capability.access_type == "mount") {
             if (
               capability.mount.fs_type &&
-              !["ext3", "ext4", "ext4dev", "xfs"].includes(
-                capability.mount.fs_type
-              )
+              !this.getFSTypes().includes(capability.mount.fs_type)
             ) {
               message = `invalid fs_type ${capability.mount.fs_type}`;
               return false;
             }
           }
 
-          if (
-            ![
-              "UNKNOWN",
-              "SINGLE_NODE_WRITER",
-              "SINGLE_NODE_SINGLE_WRITER", // added in v1.5.0
-              "SINGLE_NODE_MULTI_WRITER", // added in v1.5.0
-              "SINGLE_NODE_READER_ONLY",
-              "MULTI_NODE_READER_ONLY",
-              "MULTI_NODE_SINGLE_WRITER",
-            ].includes(capability.access_mode.mode)
-          ) {
+          if (!this.getAccessModes().includes(capability.access_mode.mode)) {
             message = `invalid access_mode, ${capability.access_mode.mode}`;
             return false;
           }
@@ -380,6 +361,11 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
       }
     }
 
+    let accessible_topology;
+    if (typeof this.getAccessibleTopology === "function") {
+      accessible_topology = await this.getAccessibleTopology();
+    }
+
     let volume = {
       // remove parent dataset info
       volume_id: row["name"].replace(new RegExp("^" + datasetName + "/"), ""),
@@ -389,6 +375,7 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
           : row["volsize"],
       content_source: volume_content_source,
       volume_context,
+      accessible_topology,
     };
 
     return volume;
@@ -402,7 +389,7 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
    */
   async getMaxZvolNameLength() {
     const driver = this;
-    const sshClient = driver.getSshClient();
+    const execClient = driver.getExecClient();
 
     let response;
     let command;
@@ -412,7 +399,7 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
     // get kernel
     command = "uname -s";
     driver.ctx.logger.verbose("uname command: %s", command);
-    response = await sshClient.exec(command);
+    response = await execClient.exec(command);
     if (response.code !== 0) {
       throw new Error("failed to run uname to determine max zvol name length");
     } else {
@@ -429,7 +416,7 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
         // get kernel_release
         command = "uname -r";
         driver.ctx.logger.verbose("uname command: %s", command);
-        response = await sshClient.exec(command);
+        response = await execClient.exec(command);
         if (response.code !== 0) {
           throw new Error(
             "failed to run uname to determine max zvol name length"
@@ -481,35 +468,35 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
        * limit the actual checks semi sanely
        * health checks should kick in an restart the pod
        * this process is 2 checks in 1
-       * - ensure basic ssh connectivity
+       * - ensure basic exec connectivity
        * - ensure csh is not the operative shell
        */
-      if (!driver.currentSSHShell || timerEnabled === false) {
-        const sshClient = this.getSshClient();
-        driver.ctx.logger.debug("performing ssh sanity check..");
-        const response = await sshClient.exec("echo $0");
-        driver.currentSSHShell = response.stdout.split("\n")[0];
+      if (!driver.currentExecShell || timerEnabled === false) {
+        const execClient = this.getExecClient();
+        driver.ctx.logger.debug("performing exec sanity check..");
+        const response = await execClient.exec("echo $0");
+        driver.currentExecShell = response.stdout.split("\n")[0];
       }
 
       // update in the background every X interval to prevent incessant checks
-      if (timerEnabled && !driver.currentSSHShellInterval) {
+      if (timerEnabled && !driver.currentExecShellInterval) {
         const intervalTime = 60000;
-        driver.currentSSHShellInterval = setInterval(async () => {
+        driver.currentExecShellInterval = setInterval(async () => {
           try {
-            driver.ctx.logger.debug("performing ssh sanity check..");
-            const sshClient = this.getSshClient();
-            const response = await sshClient.exec("echo $0");
-            driver.currentSSHShell = response.stdout.split("\n")[0];
+            driver.ctx.logger.debug("performing exec sanity check..");
+            const execClient = this.getExecClient();
+            const response = await execClient.exec("echo $0");
+            driver.currentExecShell = response.stdout.split("\n")[0];
           } catch (e) {
-            delete driver.currentSSHShell;
+            delete driver.currentExecShell;
           }
         }, intervalTime);
       }
 
-      if (driver.currentSSHShell.includes("csh")) {
+      if (driver.currentExecShell.includes("csh")) {
         throw new GrpcError(
           grpc.status.FAILED_PRECONDITION,
-          `csh is an unsupported shell, please update the default shell of your ssh user`
+          `csh is an unsupported shell, please update the default shell of your exec user`
         );
       }
 
@@ -533,7 +520,7 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
   async CreateVolume(call) {
     const driver = this;
     const driverZfsResourceType = this.getDriverZfsResourceType();
-    const sshClient = this.getSshClient();
+    const execClient = this.getExecClient();
     const zb = await this.getZetabyte();
 
     let datasetParentName = this.getVolumeParentDatasetName();
@@ -1028,7 +1015,7 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
 
         // set mode
         if (this.options.zfs.datasetPermissionsMode) {
-          command = sshClient.buildCommand("chmod", [
+          command = execClient.buildCommand("chmod", [
             this.options.zfs.datasetPermissionsMode,
             properties.mountpoint.value,
           ]);
@@ -1037,7 +1024,7 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
           }
 
           driver.ctx.logger.verbose("set permission command: %s", command);
-          response = await sshClient.exec(command);
+          response = await execClient.exec(command);
           if (response.code != 0) {
             throw new GrpcError(
               grpc.status.UNKNOWN,
@@ -1053,7 +1040,7 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
           this.options.zfs.datasetPermissionsUser ||
           this.options.zfs.datasetPermissionsGroup
         ) {
-          command = sshClient.buildCommand("chown", [
+          command = execClient.buildCommand("chown", [
             (this.options.zfs.datasetPermissionsUser
               ? this.options.zfs.datasetPermissionsUser
               : "") +
@@ -1068,7 +1055,7 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
           }
 
           driver.ctx.logger.verbose("set ownership command: %s", command);
-          response = await sshClient.exec(command);
+          response = await execClient.exec(command);
           if (response.code != 0) {
             throw new GrpcError(
               grpc.status.UNKNOWN,
@@ -1082,7 +1069,7 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
         // probably could see if ^-.*\s and split and then shell escape
         if (this.options.zfs.datasetPermissionsAcls) {
           for (const acl of this.options.zfs.datasetPermissionsAcls) {
-            command = sshClient.buildCommand("setfacl", [
+            command = execClient.buildCommand("setfacl", [
               acl,
               properties.mountpoint.value,
             ]);
@@ -1091,7 +1078,7 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
             }
 
             driver.ctx.logger.verbose("set acl command: %s", command);
-            response = await sshClient.exec(command);
+            response = await execClient.exec(command);
             if (response.code != 0) {
               throw new GrpcError(
                 grpc.status.UNKNOWN,
@@ -1137,6 +1124,11 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
     // this should give us a relatively sane way to clean up artifacts over time
     await zb.zfs.set(datasetName, { [SUCCESS_PROPERTY_NAME]: "true" });
 
+    let accessible_topology;
+    if (typeof this.getAccessibleTopology === "function") {
+      accessible_topology = await this.getAccessibleTopology();
+    }
+
     const res = {
       volume: {
         volume_id: name,
@@ -1148,6 +1140,7 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
             : 0,
         content_source: volume_content_source,
         volume_context,
+        accessible_topology,
       },
     };
 
@@ -1562,7 +1555,6 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
       }
       entries = this.ctx.cache.get(`ListVolumes:result:${uuid}`);
       if (entries) {
-        entries = JSON.parse(JSON.stringify(entries));
         entries_length = entries.length;
         entries = entries.slice(start_position, end_position);
         if (max_entries > 0 && end_position > entries_length) {
@@ -1659,10 +1651,7 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
 
     if (max_entries && entries.length > max_entries) {
       uuid = uuidv4();
-      this.ctx.cache.set(
-        `ListVolumes:result:${uuid}`,
-        JSON.parse(JSON.stringify(entries))
-      );
+      this.ctx.cache.set(`ListVolumes:result:${uuid}`, entries);
       next_token = `${uuid}:${max_entries}`;
       entries = entries.slice(0, max_entries);
     }
@@ -2366,4 +2355,4 @@ class ControllerZfsSshBaseDriver extends CsiBaseDriver {
   }
 }
 
-module.exports.ControllerZfsSshBaseDriver = ControllerZfsSshBaseDriver;
+module.exports.ControllerZfsBaseDriver = ControllerZfsBaseDriver;