+++ src/driver/factory.js
@@ -1,6 +1,10 @@
 const { FreeNASSshDriver } = require("./freenas/ssh");
 const { FreeNASApiDriver } = require("./freenas/api");
+const {
+  ControllerLocalHostpathDriver,
+} = require("./controller-local-hostpath");
 const { ControllerZfsGenericDriver } = require("./controller-zfs-generic");
+const { ControllerZfsLocalDriver } = require("./controller-zfs-local");
 const {
   ZfsLocalEphemeralInlineDriver,
 } = require("./zfs-local-ephemeral-inline");
@@ -31,12 +35,17 @@ function factory(ctx, options) {
     case "zfs-generic-nfs":
     case "zfs-generic-iscsi":
       return new ControllerZfsGenericDriver(ctx, options);
+    case "zfs-local-dataset":
+    case "zfs-local-zvol":
+      return new ControllerZfsLocalDriver(ctx, options);
     case "zfs-local-ephemeral-inline":
       return new ZfsLocalEphemeralInlineDriver(ctx, options);
     case "smb-client":
       return new ControllerSmbClientDriver(ctx, options);
     case "nfs-client":
       return new ControllerNfsClientDriver(ctx, options);
+    case "local-hostpath":
+      return new ControllerLocalHostpathDriver(ctx, options);
     case "lustre-client":
       return new ControllerLustreClientDriver(ctx, options);
     case "node-manual":