+++ CHANGELOG.md
@@ -1,6 +1,21 @@
+# v1.5.0
+
+Released 2022-02-23
+
+- massive ci overhaul
+  - add additional drivers
+  - add additional TrueNAS server versions
+  - only build `node_modules` once by using artifacts
+  - support allow/block listing specific tests
+  - better logic waiting for driver socket to appear
+- introduce `zfs-local-dataset` driver
+- introduce `zfs-local-zvol` driver
+- introduce `local-hostpath` driver
+- support manually provisioned (`node-manual`) `oneclient` volumes
+
 # v1.4.4
 
-Release 2021-12-11
+Released 2021-12-11
 
 - better adherence to expected csi behavior when volume request for less than
   minimum volume size is requested (see #137)